🪅 Message Effects & Refactors
