+++ lib/src/televerse/extensions/extensions.dart
@@ -5,3 +5,4 @@ import 'package:televerse/telegram.dart';
 import 'package:televerse/televerse.dart';
 
 part 'user_mention.dart';
+part 'message_ext.dart';
